requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.31
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
ogb>=1.3.6
torch>=2.0
